Add several US-common unit conversions
The "ounce-inches" are not uncommon torque measurements for small
electric machines, are really just 1/16 of an lbf.

Inches of water and inches of mercury are not-unusual pressure
measurements.  The values committed here are the "conventional" units,
as defined by ISO80000-3, (which are mercury at 0 deg C and water at
2 deg C).

A well-maintained set of unit conversions can be found at https://www.nist.gov/pml/special-publication-811/nist-guide-si-appendix-b-conversion-factors/nist-guide-si-appendix-b8